[FIX] website: perf - get most specific in SQL
Before this commit, we would get all the pages for the requested URL then
filter to get the most specific.
This can just be done in SQL by sorting website_id desc so the specific are
returned first.

Note that this will respect 41897f0dbab.

This will avoid having to prefetch the generic page when it is useless.

Related to #47257
task-2211013