[FIX] stock: don't write stuff in session.user_context
The stock traceability report client action inadvertently wrote
information directly in the user_context. Obviously, this could
cause issues (mainly because it added a 'active_id' key). Indeed,
as soon as this action was executed, the 'active_id' key was
always present in the context (even for actions executed in other
apps), and could cause python methods to apply on wrong records.

opw 2216829

closes odoo/odoo#47991

X-original-commit: ef06e32202af063bc1fe70d65cd76862746c62e6
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>