[FIX] stock: unlink zero quants
If you create a quant through the inventory mode of the list view and
you don't set a quantity, the quantity is null and the quant isn't
unlinked by _unlink_zero_quants.

Also consider null quantity as 0.

We don't have the issue for the reserved_quantity field since this one
is not nullable.

closes odoo/odoo#47983

Signed-off-by: Simon Lejeune (sle) <sle@openerp.com>