[FIX] website_profile: fix edit profile issue
PURPOSE

Before this commit when an administrator edit another user's profile it
was his own data that displayed for the changes.
This commit fixes this issue.

SPECIFICATION

Add a test in both controller route when displaying the edit view and
when submitting the changes that check if the edited user is the same
as the actual user.

LINKS

Task ID : 2222043
PR : #48262

closes odoo/odoo#48351

X-original-commit: 81f18bbc4d2df4ce008ce40308909270ad9ec841
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>