[FIX] point_of_sale: closing session with zero-valued invoice
The problem is described in https://github.com/odoo/odoo/issues/48550.
This is caused by reconciled receivable line in the invoice being
reconciled with another reconciled receivable line in pos_session's
move_id. The lines are already reconciled because their balance is zero.
So to avoid trying to reconcile those reconciled lines, we do not
include them in the list of lines to be reconciled.

closes odoo/odoo#48563

Signed-off-by: pimodoo <pimodoo@users.noreply.github.com>